+++ CHANGELOG.md
@@ -94,6 +94,8 @@
     ([issue #444](https://github.com/PowerShell/xPSDesiredStateConfiguration/issues/444))
 - Changes to xService
   - Added support for Group Managed Service Accounts
+- Adds new Integration tests for MSFT_xDSCWebService and removes old
+  Integration test file, MSFT_xDSCWebService.xxx.ps1.
 
 ## 8.4.0.0